backend/requirements.txt
# Raktio Backend — Python Dependencies

# --- Web framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
pydantic-settings==2.3.0

# --- LLM providers ---
anthropic==0.30.0           # Claude Sonnet — PLANNING and REPORT routes
openai==1.40.0              # DeepSeek (OpenAI-compatible) — RUNTIME route

# --- Database ---
supabase==2.5.0             # Supabase client (Postgres + Auth + Storage)
sqlalchemy==2.0.30          # ORM for local models
asyncpg==0.29.0             # Async Postgres driver

# --- Job queue ---
arq==0.25.0                 # Async Redis Queue for background jobs
redis==5.0.7

# --- File parsing ---
pypdf2==3.0.1               # PDF parsing
python-docx==1.1.0          # DOCX parsing
python-multipart==0.0.9     # File upload handling

# --- Vector search (Supabase pgvector) ---
pgvector==0.3.0             # pgvector Python client

# --- Utilities ---
python-jose[cryptography]==3.3.0  # JWT
passlib[bcrypt]==1.7.4
httpx==0.27.0               # Async HTTP client
structlog==24.2.0           # Structured logging
python-dotenv==1.0.0

# --- Dev ---
pytest==8.2.0
pytest-asyncio==0.23.0
httpx                       # Already above, used by TestClient
